refactor: rename file to blueprint_provider
