+++ src/services/document_service.py
@@ -17,9 +17,9 @@ from common.tree_node_serializer import (
     tree_node_to_dict,
     tree_node_to_ref_dict,
 )
+from common.utils.blueprint_provider import get_blueprint_provider
 from common.utils.build_complex_search import build_mongo_query
 from common.utils.delete_documents import delete_by_attribute_path, delete_document
-from common.utils.get_blueprint import get_blueprint_provider
 from common.utils.get_resolved_document_by_id import resolve_references_in_entity
 from common.utils.get_storage_recipe import (
     create_default_storage_recipe,
